app-fastapi/requirements.txt
aiofiles==0.8.0; python_version >= "3.6" and python_version < "4.0"
amqp==5.1.0; python_version >= "3.7"
anyio==3.5.0; python_version >= "3.6" and python_full_version >= "3.6.2"
asgiref==3.5.0; python_version >= "3.7"
billiard==3.6.4.0; python_version >= "3.7"
celery==5.2.6; python_version >= "3.7"
cffi==1.15.0
cfgv==3.3.1; python_full_version >= "3.6.1" and python_version >= "3.7"
click-didyoumean==0.3.0; python_full_version >= "3.6.2" and python_full_version < "4.0.0" and python_version >= "3.7"
click-plugins==1.1.1; python_version >= "3.7"
click-repl==0.2.0; python_version >= "3.7"
click==8.1.2; python_full_version >= "3.6.2" and python_full_version < "4.0.0" and python_version >= "3.7"
colorama==0.4.4; python_version >= "3.7" and python_full_version < "3.0.0" and platform_system == "Windows" or platform_system == "Windows" and python_version >= "3.7" and python_full_version >= "3.5.0"
distlib==0.3.4; python_version >= "3.7" and python_full_version < "3.0.0" or python_full_version >= "3.5.0" and python_version >= "3.7"
fastapi==0.75.1; python_full_version >= "3.6.1"
filelock==3.7.0; python_version >= "3.7" and python_full_version < "3.0.0" or python_full_version >= "3.5.0" and python_version >= "3.7"
flake8==4.0.1; python_version >= "3.6"
flower==1.0.0
h11==0.13.0; python_version >= "3.7"
humanize==4.0.0; python_version >= "3.7"
identify==2.5.0; python_version >= "3.7"
idna==3.3; python_version >= "3.6" and python_full_version >= "3.6.2"
kombu==5.2.4; python_version >= "3.7"
mccabe==0.6.1; python_version >= "3.6"
motor==2.5.1; python_full_version >= "3.5.2"
nodeenv==1.6.0; python_version >= "3.7"
pillow==9.1.0; python_version >= "3.7"
platformdirs==2.5.2; python_version >= "3.7" and python_full_version < "3.0.0" or python_full_version >= "3.5.0" and python_version >= "3.7"
pre-commit==2.19.0; python_version >= "3.7"
prometheus-client==0.14.1; python_version >= "3.6"
prompt-toolkit==3.0.29; python_full_version >= "3.6.2" and python_version >= "3.7"
pycodestyle==2.8.0; python_version >= "3.6" and python_full_version < "3.0.0" or python_full_version >= "3.5.0" and python_version >= "3.6"
pycparser==2.21; python_version >= "2.7" and python_full_version < "3.0.0" or python_full_version >= "3.4.0"
pydantic==1.9.0; python_full_version >= "3.6.1"
pyflakes==2.4.0; python_version >= "3.6" and python_full_version < "3.0.0" or python_full_version >= "3.4.0" and python_version >= "3.6"
pymongo==3.12.3; python_full_version >= "3.5.2"
python-dotenv==0.20.0; python_version >= "3.5"
python-multipart==0.0.5
pytz==2022.1; python_version >= "3.7"
pyyaml==6.0; python_version >= "3.7"
rabbitmq==0.2.0
six==1.16.0; python_version >= "3.7" and python_full_version < "3.0.0" or python_full_version >= "3.5.0" and python_version >= "3.7"
sniffio==1.2.0; python_version >= "3.6" and python_full_version >= "3.6.2"
starlette==0.17.1; python_version >= "3.6" and python_full_version >= "3.6.1"
toml==0.10.2; python_version >= "3.7" and python_full_version < "3.0.0" or python_full_version >= "3.3.0" and python_version >= "3.7"
tornado==6.1; python_version >= "3.5"
typing-extensions==4.1.1; python_version >= "3.6" and python_full_version >= "3.6.1"
uvicorn==0.17.6; python_version >= "3.7"
vine==5.0.0; python_version >= "3.7"
virtualenv==20.14.1; python_version >= "3.7" and python_full_version < "3.0.0" or python_full_version >= "3.5.0" and python_version >= "3.7"
wcwidth==0.2.5; python_full_version >= "3.6.2" and python_version >= "3.7"
